ソース構成の見直し
