--- special_file/.gitignore
@@ -1,7 +1,0 @@
-*.o
-*.ko
-*.mod.c
-*.order
-test_file
-*.cmd
-.tmp_versions
